requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
CacheControl==0.13.1
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.7
cryptography==42.0.0
distro==1.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.0
firebase-admin==6.3.0
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-books-api-wrapper==1.0.3
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
msgpack==1.0.7
oauthlib==3.2.2
openai==1.9.0
orjson==3.9.12
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
tzlocal==5.2
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
